+++ build.gradle
@@ -31,7 +31,7 @@ plugins {
 ext {
     versions = [
             spock: "1.1-groovy-2.4",
-            rmf: "0.2.0-20190516132601",
+            rmf: "0.2.0-20181206101411",
             kotlin: "1.3.10 "
     ]
 }
